% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced.R
\name{mdb_json}
\alias{mdb_json}
\title{Export Data In An MDB Table To JSON}
\usage{
mdb_json(
  path,
  table,
  date_format = "\%Y-\%m-\%d",
  time_format = "\%Y-\%m-\%d \%H:\%M:\%S",
  no_unprintable = FALSE,
  n = -1L
)
}
\arguments{
\item{path}{Path to \code{.mdb}/\code{.accdb} file.}

\item{table}{Table name.}

\item{date_format}{Equivalent to \code{-D/--date-format}.}

\item{time_format}{Equivalent to \code{-T/--time-format}.}

\item{no_unprintable}{Equivalent to \code{-U/--no-unprintable}.}

\item{n}{Optional row limit.}
}
\value{
JSON string.
}
\description{
\code{mdb-json} is a utility program distributed with MDB Tools.
It produces JSON output for the given table. Such output is suitable for
parsing in a variety of languages.
}
\examples{
db <- mdbr:::.mdb_example_nwind_path()
if (nzchar(db) && requireNamespace("jsonlite", quietly = TRUE)) {
  mdb_json(db, "Products", n = 2)
}
}
