% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native.R
\name{print.mdblist}
\alias{print.mdblist}
\title{Print Method For \code{mdblist}}
\usage{
\method{print}{mdblist}(x, ...)
}
\arguments{
\item{x}{A \code{mdblist} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Pretty printer for multi-object text outputs returned by selected
\verb{mdb_*} helpers when \code{as_list = TRUE} (default).
}
