% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{mdb}
\alias{mdb}
\title{Create an mdbr DBI Driver}
\usage{
mdb()
}
\value{
A DBI driver for '.mdb' and '.accdb' files.
}
\description{
\code{mdb()} is the canonical DBI-style constructor for connecting to Microsoft
Access '.mdb' and '.accdb' files.
}
\examples{
db <- mdb_example()
conn <- DBI::dbConnect(mdb(), dbname = db)
DBI::dbListTables(conn)
DBI::dbDisconnect(conn)
}
