% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{mdb_example}
\alias{mdb_example}
\title{Get path to mdbr example}
\usage{
mdb_example(path = "nycflights13.mdb")
}
\arguments{
\item{path}{path to the Microsoft Access file.}
}
\description{
mdbr comes bundled with a sample file from the
\href{https://github.com/tidyverse/nycflights13}{nycflights13} package in its
inst/extdata directory. This function make it easy to access.
}
