% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdb-schema.R
\name{mdb_schema}
\alias{mdb_schema}
\title{Find the column types of a table}
\usage{
mdb_schema(file, table = NULL)
}
\arguments{
\item{file}{Path to the Microsoft Access file.}

\item{table}{Name of the table, use \code{mdb_tables()}.}
}
\value{
A tibble of table columns and data types.
}
\description{
Used to determine the column types for \code{\link[=read_mdb]{read_mdb()}}.
}
