% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{mdb_schema}
\alias{mdb_schema}
\title{Column types for a table}
\usage{
mdb_schema(file, table)
}
\arguments{
\item{file}{Path to the Microsoft Access file.}

\item{table}{Name of the table, list with \code{\link[=mdb_tables]{mdb_tables()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{col_name} (character),
\code{access_type} (character), and \code{r_type} (character, one of \code{"c"},
\code{"i"}, \code{"d"}, \code{"l"}, \code{"T"}).
}
\description{
Column types for a table
}
\examples{
\dontrun{
mdb_schema(mdb_example(), "Flights")
}
}
