% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdb-tables.R
\name{mdb_tables}
\alias{mdb_tables}
\title{List tables in a database file}
\usage{
mdb_tables(file)
}
\arguments{
\item{file}{Path to the Microsoft Access file.}
}
\value{
A character vector of table names.
}
\description{
Used to find table names for \code{read_mdb()}.
}
