% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{mdb_tables}
\alias{mdb_tables}
\title{List tables in a Microsoft Access database}
\usage{
mdb_tables(file)
}
\arguments{
\item{file}{Path to the Microsoft Access file.}
}
\value{
A character vector of table names.
}
\description{
List tables in a Microsoft Access database
}
